requirements.txt
google_api_python_client==1.7.2
google_auth_oauthlib==0.4.1
google-auth==1.8.0
google-auth-httplib2==0.0.3
librosa==0.10.1
moviepy==1.0.3
mutagen==1.47.0
numpy==1.25.2
openai==0.28.1
opencv_python==4.8.1.78
Pillow==10.1.0
Pillow==10.1.0
pilmoji==2.0.4
protobuf==4.24.4
pydub==0.25.1
python-dotenv==1.0.0
Requests==2.31.0
soundfile==0.12.1
vosk==0.3.45
webrtcvad==2.0.10
wget==3.2
discord.py
